% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dera-pq-package.R
\docType{package}
\name{dera.pq-package}
\alias{dera.pq}
\alias{dera.pq-package}
\title{dera.pq: SEC DERA data sets as local Parquet files}
\description{
\code{dera.pq} downloads SEC DERA zip archives and writes the component
tab-delimited tables to Parquet files in a local \code{DATA_DIR} repository.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iangow.github.io/dera_pq/}
}

}
\author{
\strong{Maintainer}: Ian D. Gow \email{iandgow@gmail.com} (\href{https://orcid.org/0000-0002-6243-8409}{ORCID})

Authors:
\itemize{
  \item Ian D. Gow \email{iandgow@gmail.com} (\href{https://orcid.org/0000-0002-6243-8409}{ORCID})
}

}
\keyword{internal}
