% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dera_file_metadata}
\alias{dera_file_metadata}
\title{Read DERA metadata embedded in a Parquet file}
\usage{
dera_file_metadata(path)
}
\arguments{
\item{path}{Path to a Parquet file written by this package.}
}
\value{
A named list of Parquet schema metadata.
}
\description{
Read DERA metadata embedded in a Parquet file
}
